backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
class-registry==2.1.2
coverage==7.9.1
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.3
django-environ==0.12.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
filters==1.3.2
filters-django==1.0.5
flake8==7.2.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.0
pytest-cov==6.2.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
ruff==0.8.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
